Struct yukikaze::client::Client

pub struct Client<C = DefaultCfg> where
    C: Config + 'static,  { /* fields omitted */ }

HTTP Client

Methods

impl<C: Config> Client<C> where
<C::Connector as Service<Uri>>::Error: Into<Box<dyn Error + Send + Sync + 'static>>,
<C::Connector as Service<Uri>>::Future: Send + Unpin,
<C::Connector as Service<Uri>>::Response: AsyncRead + AsyncWrite + Connection + Unpin + Send,

pub fn new() -> Client<C>

Creates new instance of client with specified configuration.

Use Default if you'd like to use default config.

pub async fn request<'_>(&'_ self, __arg1: Request) -> RequestResult

Sends request, and returns response

pub async fn send<'_>(
&'_ self,
__arg1: Request
) -> Result<RequestResult, Expired<impl Future<Output = RequestResult>, C::Timer>>

Sends request and returns response. Timed version.

On timeout error it returns async_timer::Expired as Error Expired implements Future that can be used to re-spawn ongoing request again.

If request resolves in time returns Result<response::Response, hyper::Error> as Ok variant.

pub async fn send_redirect(
&'static self,
req: Request
) -> Result<RequestResult, Expired<impl Future<Output = RequestResult> + 'static, C::Timer>>

Sends request and returns response, while handling redirects. Timed version.

On timeout error it returns async_timer::Expired as Error Expired implements Future that can be used to re-spawn ongoing request again.

If request resolves in time returns Result<response::Response, hyper::Error> as Ok variant.

pub async fn redirect_request<'_>(&'_ self, __arg1: Request) -> RequestResult

Sends request and returns response, while handling redirects.