Struct yukikaze::rt::client::generated::Request

pub struct Request(pub Request);

Global request

Implements Deref and DerefMut to access regular yukikaze's request.

Methods

impl Request

pub fn request(self) -> impl Future<Output = RequestResult>

Sends request, and returns future that resolves to response

pub fn send(
self
) -> impl Future<Output = Result<RequestResult, Expired<impl Future<Output = RequestResult>, impl Oneshot>>>

Sends request and returns response. Timed version.

On timeout error it returns async_timer::Expired as Error Expired implements Future that can be used to re-spawn ongoing request again.

If request resolves in time returns Result<response::Response, hyper::Error> as Ok variant.

pub fn send_redirect(
self
) -> impl Future<Output = Result<RequestResult, Expired<impl Future<Output = RequestResult> + 'static, impl Oneshot>>>

Sends request and returns response, while handling redirects. Timed version.

On timeout error it returns async_timer::Expired as Error Expired implements Future that can be used to re-spawn ongoing request again.

If request resolves in time returns Result<response::Response, hyper::Error> as Ok variant.

pub fn redirect_request(self) -> impl Future<Output = RequestResult>

Sends request and returns response, while handling redirects.